+++ DumbSouls/src/entities/AE/AE_SnowStorm.java
@@ -21,8 +21,8 @@ public class AE_SnowStorm extends AE_Attack_Entity {
 		frames ++;
 		tickTimer ++;
 		
-		double destX = Game.mx / Game.scale;
-		double destY = Game.my / Game.scale;
+		int destX = Game.mx / Game.scale;
+		int destY = Game.my / Game.scale;
 		double startX = x + 26 - Camera.getX();
 		double startY = y + 16 - Camera.getY();
